<update> tối ưu tìm kiếm component của entity, sửa vào map offline bị lỗi

+++ game-server/src/main/java/nro/server/model/ecs/component/PositionComponent.java
@@ -15,7 +15,7 @@ public class PositionComponent extends Component {
 
     public short x, y, newX, newY;
 
-    public short teleport = 0;
+    public byte teleport = 0;
     public boolean isDirtyMove;
     public byte isOnGround, isOnGroundNew;
     public boolean wantsToChangeMap;
